feat: refactor file structure for tests
